demo/requirements.txt
django==2.0.9
django-crispy-forms>=1.6.0
django-import-export>=0.5.1
django-reversion>=2.0.0
django-formtools>=2.1
future>=0.15.2
httplib2>=0.9.2
six>=1.10.0
jinja2>=2.10
django-rest-swagger>=2.2.0
Pillow>=5.1.0
pymysql>=0.8.1
dicttoxml>=1.7.4
xmltodict>=0.11.0
djangorestframework>=3.8.2
django-filter>=2.0.0
qrcode>=6.0
django-cors-headers>=2.4.0
cryptography>=2.3.1
pycryptodomex>=3.6.6
pycryptodome>=3.6.6